refactor(Twitter): Move redirect BM to settings patch dependancy
